Refactor visualization and interpolation modules
Streamlined the codebase by centralizing curve plotting and interpolation utilities within the geometrics module. Removed redundant implementations, improved modularity, and standardized API references for easier maintenance and extensibility. (ref. #116)
