+++ examples/examples_curves/src/bin/d2_curve.rs
@@ -1,11 +1,11 @@
 use optionstratlib::curves::visualization::Plottable;
 use optionstratlib::curves::{Curve, Point2D};
+use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 use optionstratlib::greeks::d2;
 use optionstratlib::utils::setup_logger;
 use optionstratlib::{pos, Positive};
 use rust_decimal_macros::dec;
 use std::error::Error;
-use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 
 fn main() -> Result<(), Box<dyn Error>> {
     setup_logger();
@@ -14,7 +14,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         t_end: dec!(100),
         steps: 100,
     };
-    
+
     let parametric_curve = Curve::construct(ConstructionMethod::Parametric {
         f: Box::new(|t| {
             let strike = Positive::new_decimal(t).unwrap();
