+++ examples/examples_curves/src/bin/gamma_curve.rs
@@ -1,12 +1,12 @@
 use optionstratlib::curves::visualization::Plottable;
 use optionstratlib::curves::{Curve, Point2D};
+use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 use optionstratlib::greeks::Greeks;
 use optionstratlib::utils::setup_logger;
 use optionstratlib::{pos, ExpirationDate, OptionStyle, OptionType, Options, Positive, Side};
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use std::error::Error;
-use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 
 fn get_option(strike: &Positive) -> Options {
     Options::new(
@@ -31,7 +31,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         t_end: dec!(90),
         steps: 100,
     };
-    
+
     let parametric_curve = Curve::construct(ConstructionMethod::Parametric {
         f: Box::new(|t| {
             let option = get_option(&Positive::new_decimal(t).unwrap());
@@ -39,7 +39,7 @@ fn main() -> Result<(), Box<dyn Error>> {
             let point = Point2D::new(t, value);
             Ok(point)
         }),
-        params: params.clone(), 
+        params: params.clone(),
     })?;
 
     parametric_curve
