+++ examples/examples_curves/src/bin/gamma_volatility_vector_curve.rs
@@ -1,12 +1,12 @@
 use optionstratlib::curves::visualization::Plottable;
 use optionstratlib::curves::{Curve, Point2D};
+use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 use optionstratlib::greeks::Greeks;
 use optionstratlib::utils::setup_logger;
 use optionstratlib::{pos, ExpirationDate, OptionStyle, OptionType, Options, Positive, Side};
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use std::error::Error;
-use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 
 fn get_option(strike: &Positive, volatility: &Positive) -> Options {
     Options::new(
@@ -32,7 +32,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         t_end: dec!(78.0),
         steps: 100,
     };
-    
+
     let vol_20_curve = Curve::construct(ConstructionMethod::Parametric {
         f: Box::new(|t| {
             let option = get_option(&Positive::new_decimal(t).unwrap(), &pos!(0.20));
