+++ examples/examples_curves/src/bin/parametric_vector_curve.rs
@@ -1,9 +1,9 @@
 use optionstratlib::curves::visualization::Plottable;
 use optionstratlib::curves::{Curve, Point2D};
+use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 use optionstratlib::utils::setup_logger;
 use rust_decimal::{Decimal, MathematicalOps};
 use std::error::Error;
-use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 
 fn main() -> Result<(), Box<dyn Error>> {
     setup_logger();
