+++ examples/examples_curves/src/bin/theta_curve.rs
@@ -1,12 +1,12 @@
 use optionstratlib::curves::visualization::Plottable;
 use optionstratlib::curves::{Curve, Point2D};
+use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 use optionstratlib::greeks::Greeks;
 use optionstratlib::utils::setup_logger;
 use optionstratlib::{pos, ExpirationDate, OptionStyle, OptionType, Options, Positive, Side};
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use std::error::Error;
-use optionstratlib::geometrics::{ConstructionMethod, ConstructionParams, GeometricObject};
 
 fn get_option(strike: &Positive) -> Options {
     Options::new(
