+++ src/curves/visualization/plotters.rs
@@ -15,8 +15,7 @@
 //! use std::path::Path;
 //! use rust_decimal::Decimal;
 //! use optionstratlib::curves::{Curve, Point2D};
-//! use optionstratlib::curves::visualization::Plottable;
-//! use optionstratlib::geometrics::GeometricObject;
+//! use optionstratlib::geometrics::{GeometricObject, Plottable};
 //!
 //! let curve = Curve::from_vector(vec![
 //!             Point2D::new(Decimal::ZERO, Decimal::ZERO), // p11
@@ -48,131 +47,11 @@
 
 use crate::curves::Curve;
 use crate::error::CurvesError;
+use crate::geometrics::{PlotBuilder, PlotBuilderExt, PlotOptions, Plottable};
 #[cfg(not(target_arch = "wasm32"))]
 use num_traits::ToPrimitive;
 use plotters::prelude::*;
 use std::path::Path;
-
-/// Plot configuration options
-#[derive(Clone, Debug)]
-pub struct PlotOptions {
-    /// Plot title
-    pub title: Option<String>,
-    /// X-axis label
-    pub x_label: Option<String>,
-    /// Y-axis label
-    pub y_label: Option<String>,
-    /// Line colors for each curve
-    pub line_colors: Option<Vec<RGBColor>>,
-    /// Line width
-    pub line_width: u32,
-    /// Background color
-    pub background_color: RGBColor,
-    /// Plot width
-    pub width: u32,
-    /// Plot height
-    pub height: u32,
-    /// Curve names
-    pub curve_name: Option<Vec<String>>,
-}
-
-#[allow(dead_code)]
-impl PlotOptions {
-    /// Default color palette for multiple curves
-    fn default_colors() -> Vec<RGBColor> {
-        vec![
-            RGBColor(0, 0, 255),   // Blue
-            RGBColor(255, 0, 0),   // Red
-            RGBColor(0, 255, 0),   // Green
-            RGBColor(255, 165, 0), // Orange
-            RGBColor(128, 0, 128), // Purple
-        ]
-    }
-}
-
-impl Default for PlotOptions {
-    fn default() -> Self {
-        PlotOptions {
-            title: None,
-            x_label: None,
-            y_label: None,
-            line_colors: None,
-            line_width: 2,
-            background_color: RGBColor(255, 255, 255), // White
-            width: 800,                                // Dimensión por defecto
-            height: 600,                               // Dimensión por defecto
-            curve_name: None,
-        }
-    }
-}
-
-/// Trait for plotting curves
-pub trait Plottable {
-    /// Creates a plot builder
-    fn plot(&self) -> PlotBuilder<Self>
-    where
-        Self: Sized;
-}
-
-/// Plot Builder for configurable curve visualization
-#[allow(dead_code)]
-pub struct PlotBuilder<T: Plottable> {
-    /// Data to be plotted
-    data: T,
-    /// Plot configuration options
-    options: PlotOptions,
-}
-
-impl<T: Plottable> PlotBuilder<T> {
-    /// Set plot title
-    pub fn title(mut self, title: impl Into<String>) -> Self {
-        self.options.title = Some(title.into());
-        self
-    }
-
-    /// Set x-axis label
-    pub fn x_label(mut self, label: impl Into<String>) -> Self {
-        self.options.x_label = Some(label.into());
-        self
-    }
-
-    /// Set y-axis label
-    pub fn y_label(mut self, label: impl Into<String>) -> Self {
-        self.options.y_label = Some(label.into());
-        self
-    }
-
-    pub fn curve_name(mut self, label: Vec<String>) -> Self {
-        self.options.curve_name = Some(label);
-        self
-    }
-
-    /// Set line colors
-    pub fn line_colors(mut self, colors: Vec<RGBColor>) -> Self {
-        self.options.line_colors = Some(colors);
-        self
-    }
-
-    /// Set line width
-    pub fn line_width(mut self, width: u32) -> Self {
-        self.options.line_width = width;
-        self
-    }
-
-    /// Set plot dimensions
-    pub fn dimensions(mut self, width: u32, height: u32) -> Self {
-        self.options.width = width;
-        self.options.height = height;
-        self
-    }
-
-    pub fn save(self, path: impl AsRef<Path>) -> Result<(), CurvesError>
-    where
-        Self: PlotBuilderExt<T>,
-    {
-        PlotBuilderExt::save(self, path)
-    }
-}
 
 /// Plottable implementation for single Curve
 impl Plottable for Curve {
@@ -263,12 +142,6 @@ impl Plottable for Vec<Curve> {
             options: PlotOptions::default(),
         }
     }
-}
-
-/// Plotting extension methods
-pub trait PlotBuilderExt<T: Plottable> {
-    /// Save plot to file
-    fn save(self, path: impl AsRef<Path>) -> Result<(), CurvesError>;
 }
 
 /// Plotting implementation for single Curve
@@ -551,9 +424,9 @@ impl PlotBuilderExt<Vec<Curve>> for PlotBuilder<Vec<Curve>> {
 mod tests {
     use super::*;
     use crate::curves::Point2D;
+    use crate::geometrics::GeometricObject;
     use rust_decimal_macros::dec;
     use std::fs;
-    use crate::geometrics::GeometricObject;
 
     fn cleanup_image(filename: &str) {
         if Path::new(filename).exists() {
@@ -582,7 +455,7 @@ mod tests {
         (
             Curve::from_vector(points1),
             Curve::from_vector(points2),
-            Curve::from_vector(points3)
+            Curve::from_vector(points3),
         )
     }
 
@@ -591,11 +464,7 @@ mod tests {
         let p1 = Point2D::new(dec!(0.0), dec!(0.0));
         let p2 = Point2D::new(dec!(1.0), dec!(1.0));
         let p3 = Point2D::new(dec!(2.0), dec!(4.0));
-        let points = vec![
-            &p1,
-            &p2,
-            &p3,
-        ];
+        let points = vec![&p1, &p2, &p3];
         let curve = Curve::from_vector(points);
 
         // Plot single curve
@@ -622,7 +491,7 @@ mod tests {
 
         let points1 = vec![&p1_1, &p1_2, &p1_3];
         let points2 = vec![&p2_1, &p2_2, &p2_3];
-        
+
         let curve1 = Curve::from_vector(points1);
         let curve2 = Curve::from_vector(points2);
