+++ src/geometrics/mod.rs
@@ -1,21 +1,23 @@
 /******************************************************************************
-    Author: Joaquín Béjar García
-    Email: jb@taunais.com 
-    Date: 21/1/25
- ******************************************************************************/
-mod utils;
-mod interpolation;
+   Author: Joaquín Béjar García
+   Email: jb@taunais.com
+   Date: 21/1/25
+******************************************************************************/
 mod construction;
+mod interpolation;
+mod utils;
 
 mod operations;
+mod visualization;
 
-pub use utils::{GeometricObject, Len};
-pub use interpolation::traits::Interpolate;
-pub use interpolation::linear::LinearInterpolation;
+pub use construction::{ConstructionMethod, ConstructionParams};
 pub use interpolation::bilinear::BiLinearInterpolation;
 pub use interpolation::cubic::CubicInterpolation;
+pub use interpolation::linear::LinearInterpolation;
 pub use interpolation::spline::SplineInterpolation;
-pub use interpolation::types::InterpolationType;
-pub use construction::{ConstructionMethod, ConstructionParams};
-pub use operations::{MergeOperation, CurveArithmetic};
 pub use interpolation::traits::HasX;
+pub use interpolation::traits::Interpolate;
+pub use interpolation::types::InterpolationType;
+pub use operations::{CurveArithmetic, MergeOperation};
+pub use utils::{GeometricObject, Len};
+pub use visualization::{PlotBuilder, PlotBuilderExt, PlotOptions, Plottable};
