+++ src/geometrics/visualization/mod.rs
@@ -3,10 +3,6 @@
    Email: jb@taunais.com
    Date: 21/1/25
 ******************************************************************************/
+mod plotters;
 
-pub(crate) mod bilinear;
-pub(crate) mod cubic;
-pub(crate) mod linear;
-pub(crate) mod spline;
-pub(crate) mod traits;
-pub(crate) mod types;
+pub use plotters::{PlotBuilder, PlotBuilderExt, PlotOptions, Plottable};
