+++ src/lib.rs
@@ -705,10 +705,10 @@ pub mod risk;
 pub mod simulation;
 pub mod strategies;
 // pub mod surfaces;
+pub mod geometrics;
 pub mod utils;
 pub mod visualization;
 pub mod volatility;
-pub mod geometrics;
 
 pub use model::positive::Positive;
 pub use model::types::{ExpirationDate, OptionStyle, OptionType, Side};
